Se creo un modelo llamado importacion y se mudaron las clases a el desde convenio
