+++ precious/convenio/admin.py
@@ -9,12 +9,9 @@ from .models import (
     CodigoPractica,
     DetalleArancel,
     DetalleCodigo,
-    ImportarPracticas,
-    ImportarHomologacion,
     Usuario,
     TipoPractica,
 )
-import pdb
 
 
 # Define an inline admin descriptor for Employee model
@@ -190,36 +187,3 @@ class DetalleCodigoAdmin(admin.ModelAdmin):
     list_filter = (
         'convenio',
     )
-
-
-@admin.register(ImportarPracticas)
-class ImportarPracticasAdmin(admin.ModelAdmin):
-    actions_on_bottom = True
-    list_display = (
-        'archivo',
-        'prestador',
-    )
-    exclude = ('creator', 'updater')
-
-    def save_model(self, request, obj, form, change):
-        if getattr(obj, 'creator', None) is None:
-            obj.creator = request.user
-        obj.updater = request.user
-        obj.save()
-
-
-@admin.register(ImportarHomologacion)
-class ImportarHomologacionAdmin(admin.ModelAdmin):
-    actions_on_bottom = True
-    list_display = (
-        'archivo',
-        'prestador',
-    )
-    exclude = ('creator', 'updater')
-
-    def save_model(self, request, obj, form, change):
-        if getattr(obj, 'creator', None) is None:
-            obj.creator = request.user
-        obj.updater = request.user
-        obj.save()
-
