+++ precious/precious/settings.py
@@ -38,6 +38,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'convenio',
+    'importacion',
 ]
 
 MIDDLEWARE = [
